Admin cannot delete himself

+++ DeleteUser_admin.php
@@ -3,6 +3,8 @@ session_start();
 
 require("krunkerideaconn.php");
 
+$user_id = $_SESSION["userid"];
+
 if(!isset($_SESSION['role'])){
   header("Location: index.php");
   exit;
@@ -16,7 +18,7 @@ else{
 }
 $id = $_GET['id'];
 
-$sql = "DELETE FROM `user_tbl` WHERE UserId =$id";
+$sql = "DELETE FROM `user_tbl` WHERE UserId =$id AND NOT UserId='$user_id' ";
 $result =mysqli_query($dbconn, $sql);
 
   if($result){
